organize BinaryReader/BinaryWriter extensions
